feat: add model
